PCS-59/sanjeevni-fastapi/requirements.txt
fastapi
uvicorn
python-dotenv
requests
reportlab
pydantic
python-multipart
google-genai
aiohttp
gtts
pyttsx3